MB-65392 Add Graph Building status and update progress
/getIndexStatus endpoint will return new graph building status for index
when bhive graph build is in progress and also report the progress for
each graph build.

+++ secondary/indexer/storage_manager.go
@@ -1817,7 +1817,7 @@ func (s *storageMgr) handleStats(cmd Message) {
 				idxStats.insertBytes.Set(st.Stats.InsertBytes)
 				idxStats.deleteBytes.Set(st.Stats.DeleteBytes)
 
-				idxStats.graphBuildProgress.Set(float64(st.Stats.GraphBuildProgress))
+				idxStats.graphBuildProgress.Set(int64(st.Stats.GraphBuildProgress))
 
 				// compute mutation rate
 				now := time.Now().UnixNano()
